bugfix: iOS UFG_ClearArea error masking VHS extraction errors in the SDK.
The issue was this: if an error occurred during VHS taking, we would still try to do cleanup by clicking the UFG_ClearArea. However, some errors may cause this label to not be available - in which case we would throw an error about the UFG_ClearArea instead of the original error.